\name{StatComp21076-package}
\alias{StatComp21076-package}
\docType{package}
\title{
  SVT and SVP for 'Statistical Computing' course
}
\description{
  SVT and SVP function are included in this package for illustration.
}
\author{
Qingguo Wu

Qingguo Wu <sa21204165@mail.ustc.edu.cn>
}
\keyword{ package }
