% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp.R
\name{SVT}
\alias{SVT}
\title{SVT}
\usage{
SVT(M, step_size, tolerance, lamda, k_max)
}
\arguments{
\item{M}{sparse matrix}

\item{step_size}{step size}

\item{tolerance}{tolerance}

\item{lamda}{tunning parameter}

\item{k_max}{maximum number of iterations}
}
\value{
complete matrix
}
\description{
A Singular Value Thresholding Algorithm for Matrix Completion
}
\examples{
\dontrun{
M <- matrix(c(0,1,0,1,0,1,0,1,0),3,3)
f <- SVT(M,1,0.9,1,10000)
}
}
